// Frontend/src/interfaces/user.ts
export interface User {
  id: number;
  fullName: string;
  email: string;
  userName: string;
  password: string;
  address: string;
  phone: string;
  birthDate: string;
  role: boolean;
}
export interface listUser {
  id: number;
  fullName: string;
  email: string;
  userName: string;
  password: string;
  address: string;
  phone: string;
  birthDate: string;
  role: boolean;
}